requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asyncpg==0.26.0
attrs==22.1.0
click==8.1.3
fastapi==0.79.1
graphql-core==3.2.1
h11==0.13.0
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.2
Pygments==2.13.0
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.1.2
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.2.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
strawberry-graphql==0.126.2
tomli==2.0.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.2
websockets==10.3
